refactor: Firebaseインスタンスの初期化を別ファイルに切り出し
