+++ frontend/pages/sign-in.tsx
@@ -4,16 +4,7 @@ import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth";
 import firebase from "firebase/compat/app";
 import "firebase/compat/auth";
 
-// Configure Firebase.
-const config = {
-  apiKey: "AIzaSyCoOj9RP4KyGBpJttb2rf-965A1Q1HYTI4",
-  authDomain: "nest-api-server.firebaseapp.com",
-  projectId: "nest-api-server",
-  storageBucket: "nest-api-server.appspot.com",
-  messagingSenderId: "992523588163",
-  appId: "1:992523588163:web:ae8b108155461a620efaf6",
-};
-firebase.initializeApp(config);
+import "../utils/firebase/init";
 
 // Configure FirebaseUI.
 const uiConfig = {
